💄Move CustomerFileter, PersonFilter, and SelectProjectDialog to shared Components folder
