+++ Satori/Pages/Components/PersonFilter.razor.cs
@@ -1,8 +1,9 @@
 ﻿using System.Diagnostics.CodeAnalysis;
 using Microsoft.AspNetCore.Components;
 using Satori.AppServices.ViewModels;
+using Satori.Pages.SprintBoard;
 
-namespace Satori.Pages.SprintBoard;
+namespace Satori.Pages.Components;
 
 public partial class PersonFilter
 {
