+++ Satori/Pages/Components/SelectProjectDialog.razor.css
