+++ Satori/Pages/SprintBoard/ActionItemView.razor
@@ -3,6 +3,7 @@
 @using Satori.AppServices.ViewModels.PullRequests.ActionItems
 @using Satori.AppServices.ViewModels.WorkItems.ActionItems
 @using Satori.Kimai
+@using Satori.Pages.Components
 
 @inject WorkItemUpdateService WorkItemUpdateService
 @inject IJSRuntime JsRuntime
