+++ Satori/Pages/SprintBoard/ActionItemView.razor.cs
@@ -8,6 +8,7 @@ using Satori.AppServices.ViewModels.PullRequests.ActionItems;
 using Satori.AppServices.ViewModels.WorkItems;
 using Satori.AppServices.ViewModels.WorkItems.ActionItems;
 using Satori.Kimai.ViewModels;
+using Satori.Pages.Components;
 
 namespace Satori.Pages.SprintBoard;
