+++ Satori/Pages/SprintBoard/SprintBoards.razor
@@ -5,6 +5,7 @@
 @using Satori.AppServices.Models
 @using Satori.AppServices.Services.Abstractions
 @using Satori.AppServices.ViewModels.WorkItems
+@using Satori.Pages.Components
 @using Toolbelt.Blazor.HotKeys2
 @attribute [StreamRendering]
 @inject NavigationManager NavigationManager
