+++ Satori/Pages/SprintBoard/SprintBoards.razor.cs
@@ -12,6 +12,7 @@ using Satori.AppServices.ViewModels.Sprints;
 using Satori.AppServices.ViewModels.WorkItems;
 using Satori.AppServices.ViewModels.WorkItems.ActionItems;
 using Satori.Kimai.ViewModels;
+using Satori.Pages.Components;
 using Satori.Utilities;
 using Toolbelt.Blazor.HotKeys2;
